Create Desktop notifier Application Using python
Well this is one of the simpler and easier way to make your own desktop notifier  in just few minutes. They will provide you a simple pop up messages on your screen.
Basically,Desktop notifier is the action of notifying someone or something. It’s main purpose behind a notification is to inform about an event and encourage him to take action.
Notifications help people to remember things. It is a small piece of text which appears on the desktop or mobile screen to inform the user about the updates or any other important pieces of Information.
Other areas where we can use this approach are:
1. Set daily tracker for COVID status
2. Daily notification to take medicine.
3. Hourly notification to drink water.
and many more, it’s completely up to you how to use this application.
Desktop notifications create a popup message each time an inbound call is received. Popup messages provide both a visual and audio notification of an inbound call and appear even when 
the browser is minimized. It is a simple Python script which parses the news headlines available in XML format. Save this Python script as topnews.py (as we import it by this name in our desktop notifier app). Now, in order to create a desktop notifier, you need to install a third party Python module, notify2.